Add files via upload
IPC as a statute has survived and flourished over the last 160 years, which speaks volumes about its effectiveness as a penal code of high stature. However, over these years it has not been able to shed away some of its provisions that reek of colonialism e.g. sedition. The Malimath Committee report while advocating criminal justice reforms has provided the Parliament with an opportunity to revamp the Code and other criminal laws. It has been 17 years since the submission of the report and no concrete steps have been taken in this respect. It is about time that the legislature steps in to make the Code more in sync with modern times than with the times of British colonialism. It does not reflect well on the legislature when the apex court intervenes and strikes off laws because it is the job of the legislature to do so in the first place.